[onchaindiscovery] move setup_onchain_discovery into onchain_discovery crate
